+++ cpp/src/wrapper/knowhere/wrapper_log.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <easylogging++.h>
+#include "utils/easylogging++.h"
 
 namespace zilliz {
 namespace milvus {