+++ cpp/CHANGELOG.md
@@ -7,6 +7,7 @@ Please mark all change in change log and use the ticket from JIRA.
 ## Bug
 
 ## Improvement
+- MS-552 - Add and change the easylogging library
 
 ## New Feature